requirements.txt
streamlit-chatbot-webapp/streamlit-chatbot-webapp/requirements.txt

streamlit
langchain
python-dotenv
torch
transformers
chromadb